+++ boards/raspberrypi/rpi_5/rpi_5.dts
@@ -23,6 +23,7 @@
 		zephyr,sram = &sram0;
 		zephyr,console = &uart10;
 		zephyr,shell-uart = &uart10;
+		zephyr,pcie-controller = &pcie2;
 	};
 
 	leds {
