+++ boards/raspberrypi/rpi_5/rpi_5_defconfig
@@ -10,3 +10,5 @@ CONFIG_CONSOLE=y
 CONFIG_UART_CONSOLE=y
 
 CONFIG_MFD=y
+CONFIG_PCIE=y
+CONFIG_PCIE_CONTROLLER=y
