+++ dts/arm64/broadcom/bcm2712.dtsi
@@ -92,46 +92,63 @@
 		#address-cells = <2>;
 		#size-cells = <1>;
 
-		rp1@1000120000 {
-			compatible = "raspberrypi,rp1";
+		pcie1: pcie@1000110000 {
+			compatible = "brcm,brcmstb-pcie";
+			reg = <0x10 0x110000 0x9310>;
+			#address-cells = <2>;
+			#size-cells = <1>;
+			status = "okay";
+		};
+
+		pcie2: pcie@1000120000 {
+			compatible = "brcm,brcmstb-pcie";
 			reg = <0x10 0x120000 0x9310>;
 			#address-cells = <2>;
 			#size-cells = <1>;
 			status = "okay";
 
-			gpio0: gpio@1f000d0000 {
-				compatible = "simple-bus";
-				reg = <0x1f 0xd0000 0x30000>;
-				#address-cells = <1>;
-				#size-cells = <0>;
+			rp1 {
+				compatible = "raspberrypi,rp1";
+				//reg = <0x10 0x120000 0x9310>;
+				//#address-cells = <2>;
+				//#size-cells = <1>;
+				status = "okay";
 
-				gpio0_0: gpio@0 {
-					compatible = "raspberrypi,rp1-gpio";
-					reg = <0x0 0x10000 0x20000>;
-					#gpio-cells = <2>;
-					gpio-controller;
-					ngpios = <28>;
-					status = "disabled";
-				};
+				gpio0: gpio@1f000d0000 {
+					compatible = "simple-bus";
+					reg = <0x1f 0xd0000 0x30000>;
+					#address-cells = <1>;
+					#size-cells = <0>;
 
-				gpio0_1: gpio@4000 {
-					compatible = "raspberrypi,rp1-gpio";
-					reg = <0x4000 0x14000 0x24000>;
-					#gpio-cells = <2>;
-					gpio-controller;
-					ngpios = <6>;
-					status = "disabled";
-				};
+					gpio0_0: gpio@0 {
+						compatible = "raspberrypi,rp1-gpio";
+						reg = <0x0 0x10000 0x20000>;
+						#gpio-cells = <2>;
+						gpio-controller;
+						ngpios = <28>;
+						status = "disabled";
+					};
 
-				gpio0_2: gpio@8000 {
-					compatible = "raspberrypi,rp1-gpio";
-					reg = <0x8000 0x18000 0x28000>;
-					#gpio-cells = <2>;
-					gpio-controller;
-					ngpios = <20>;
-					status = "disabled";
+					gpio0_1: gpio@4000 {
+						compatible = "raspberrypi,rp1-gpio";
+						reg = <0x4000 0x14000 0x24000>;
+						#gpio-cells = <2>;
+						gpio-controller;
+						ngpios = <6>;
+						status = "disabled";
+					};
+
+					gpio0_2: gpio@8000 {
+						compatible = "raspberrypi,rp1-gpio";
+						reg = <0x8000 0x18000 0x28000>;
+						#gpio-cells = <2>;
+						gpio-controller;
+						ngpios = <20>;
+						status = "disabled";
+					};
 				};
 			};
 		};
+
 	};
 };
