+++ dts/bindings/mfd/raspberrypi,rp1.yaml
@@ -9,4 +9,4 @@ include: base.yaml
 
 properties:
   reg:
-    required: true
+    required: false
