+++ dts/bindings/pcie/controller/brcm,brcmstb-pcie.yaml
@@ -1,0 +1,12 @@
+# Copyright 2024 TSNLab, Inc.
+# SPDX-License-Identifier: Apache-2.0
+
+description: BRCMSTB PCIe controller
+
+compatible: "brcm,brcmstb-pcie"
+
+include: [base.yaml, pcie-controller.yaml]
+
+properties:
+  reg:
+    required: true
